requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-filter==2.1.0
django-haystack==2.8.1
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
jieba==0.39
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
Pygments==2.4.2
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
Whoosh==2.7.4
